Update/client ressources

+++ src/Controller/UserController.php
@@ -8,6 +8,7 @@ use App\Entity\User;
 use App\Pagination\PaginationFactory;
 use App\Repository\ClientRepository;
 use App\Repository\UserRepository;
+use DateTime;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
@@ -54,7 +55,7 @@ class UserController extends AbstractController
     #[Route('', name: 'user_index', methods: ['GET'])]
     public function listAction(UserRepository $userRepository, Request $request): Response
     {
-        $qb = $userRepository->findAllQueryBuilder();
+        $qb = $userRepository->findUserByClient($this->getUser()->getId());
 
         $paginatedCollection = $this->paginationFactory
             ->createCollection($qb, $request, 'user_index');
@@ -63,11 +64,11 @@ class UserController extends AbstractController
     }
 
     /**
-     * Create a new Bilemo customer
+     * Create a new Bilemo customer linked to the current logged user
      *
      * @OA\Response(
      *     response=201,
-     *     description="Create a new customer with the data sent in the request",
+     *     description="Create a new customer with the data sent in the request. The customer is linked in a manyToOne relation to it's brand.",
      *     @OA\JsonContent(
      *      type="array",
      *      @OA\Items(ref=@Model(type=User::class, groups={"user:read"}))
@@ -78,7 +79,7 @@ class UserController extends AbstractController
      * @Security(name="Bearer")
      */
     #[Route('', name: 'user_new', methods: ['POST'])]
-    public function createAction(Request $request, ClientRepository $clientRepository): Response
+    public function createAction(Request $request): Response
     {
         try {
             $user = $this->serializer->deserialize($request->getContent(), User::class, 'json');
@@ -91,9 +92,8 @@ class UserController extends AbstractController
 
             throw new ApiProblemException($apiProblem);
         }
-        // Replace with current client later
-        $defaultClient = $clientRepository->findOneBy(['brand' => 'FSR']);
-        $user->setClient($defaultClient);
+
+        $user->setClient($this->getUser());
 
         $errors = $this->validator->validate($user);
 
@@ -130,7 +130,15 @@ class UserController extends AbstractController
     #[Route('/{id}', name: 'user_show', methods: ['GET'])]
     public function showAction(User $user): Response
     {
-        return $this->json($user, Response::HTTP_OK, [], ['groups' => 'user:read']);
+        if ($user->getClient()->getUsername() === $this->getUser()->getUsername()) {
+            return $this->json($user, Response::HTTP_OK, [], ['groups' => 'user:read']);
+        } else {
+            $apiProblem = new ApiProblem(
+                400,
+                ApiProblem::TYPE_FORBIDDEN_RESSOURCE,
+            );
+            throw new ApiProblemException($apiProblem);
+        }
     }
 
     /**
@@ -160,7 +168,7 @@ class UserController extends AbstractController
             $userRequest = $this->serializer->deserialize($request->getContent(), User::class, 'json');
         } catch (NotEncodableValueException $e) {
             $apiProblem = new ApiProblem(
-                400,
+                403,
                 ApiProblem::TYPE_INVALID_REQUEST_BODY_FORMAT,
             );
             $apiProblem->set('errors', $e);
@@ -168,17 +176,25 @@ class UserController extends AbstractController
             throw new ApiProblemException($apiProblem);
         }
 
-        $errors = $this->validator->validate($userRequest);
+        if ($userExist->getClient()->getUsername() === $this->getUser()->getUsername()) {
+            $errors = $this->validator->validate($userRequest);
 
-        if (count($errors) > 0) {
-            return $this->throwApiProblemValidationException($errors);
+            if (count($errors) > 0) {
+                return $this->throwApiProblemValidationException($errors);
+            }
+
+            $userExist->setFirstName($userRequest->getFirstName());
+            $userExist->setLastName($userRequest->getLastName());
+            $userExist->setPhoneNumber($userRequest->getPhoneNumber());
+            $userExist->setEmailAddress($userRequest->getEmailAddress());
+            $userExist->setUpdatedAt(new DateTime());
+
+            $this->entityManager->flush();
+
+            return $this->json($userExist, Response::HTTP_OK, [], ['groups' => 'user:read']);
+        } else {
+            $this->throwApiProblemForbiddenRessourceException();
         }
-
-        $userExist->setEmailAddress($userRequest->getEmailAddress());
-
-        $this->entityManager->flush();
-
-        return $this->json($userRequest, Response::HTTP_OK, [], ['groups' => 'user:read']);
     }
 
     /**
@@ -204,30 +220,36 @@ class UserController extends AbstractController
     #[Route('/{id}', name: 'user_edit_patch', methods: ['PATCH'])]
     public function editPatchAction(Request $request, User $user): Response
     {
-        try {
-            $user = $this->serializer->deserialize($request->getContent(), User::class, 'json', [
-                AbstractNormalizer::OBJECT_TO_POPULATE => $user,
-                'groups' => 'user:write'
-            ]);
-        } catch (NotEncodableValueException $e) {
-            $apiProblem = new ApiProblem(
-                400,
-                ApiProblem::TYPE_INVALID_REQUEST_BODY_FORMAT,
-            );
-            $apiProblem->set('errors', $e);
+        if ($user->getClient()->getUsername() === $this->getUser()->getUsername()) {
+            try {
+                $user = $this->serializer->deserialize($request->getContent(), User::class, 'json', [
+                    AbstractNormalizer::OBJECT_TO_POPULATE => $user,
+                    'groups' => 'user:write'
+                ]);
+            } catch (NotEncodableValueException $e) {
+                $apiProblem = new ApiProblem(
+                    400,
+                    ApiProblem::TYPE_INVALID_REQUEST_BODY_FORMAT,
+                );
+                $apiProblem->set('errors', $e);
 
-            throw new ApiProblemException($apiProblem);
+                throw new ApiProblemException($apiProblem);
+            }
+
+            $errors = $this->validator->validate($user);
+
+            if (count($errors) > 0) {
+                $this->throwApiProblemValidationException($errors);
+            }
+
+            $user->setUpdatedAt(new DateTime());
+
+            $this->entityManager->flush();
+
+            return $this->json($user, Response::HTTP_OK, [], ['groups' => 'user:read']);
+        } else {
+            $this->throwApiProblemForbiddenRessourceException();
         }
-
-        $errors = $this->validator->validate($user);
-
-        if (count($errors) > 0) {
-            $this->throwApiProblemValidationException($errors);
-        }
-
-        $this->entityManager->flush();
-
-        return $this->json($user, Response::HTTP_OK, [], ['groups' => 'user:read']);
     }
 
     /**
@@ -249,10 +271,14 @@ class UserController extends AbstractController
     #[Route('/{id}', name: 'user_delete', methods: ['DELETE'])]
     public function deleteAction(User $user): Response
     {
-        $this->entityManager->remove($user);
-        $this->entityManager->flush();
+        if ($user->getClient()->getUsername() === $this->getUser()->getUsername()) {
+            $this->entityManager->remove($user);
+            $this->entityManager->flush();
 
-        return new Response(null, Response::HTTP_NO_CONTENT);
+            return new Response(null, Response::HTTP_NO_CONTENT);
+        } else {
+            $this->throwApiProblemForbiddenRessourceException();
+        }
     }
 
     private function throwApiProblemValidationException($errors): Response
@@ -270,6 +296,15 @@ class UserController extends AbstractController
 
         $apiProblem->set('errors', $errorsMessage);
 
+        throw new ApiProblemException($apiProblem);
+    }
+
+    private function throwApiProblemForbiddenRessourceException()
+    {
+        $apiProblem = new ApiProblem(
+            403,
+            ApiProblem::TYPE_FORBIDDEN_RESSOURCE,
+        );
         throw new ApiProblemException($apiProblem);
     }
 }
